Created plane talker node and comms listener node
