+++ log/build_2023-05-10_22-42-45/autonomous/stdout_stderr.log
@@ -1,0 +1,22 @@
+running egg_info
+writing ../../build/autonomous/autonomous.egg-info/PKG-INFO
+writing dependency_links to ../../build/autonomous/autonomous.egg-info/dependency_links.txt
+writing entry points to ../../build/autonomous/autonomous.egg-info/entry_points.txt
+writing requirements to ../../build/autonomous/autonomous.egg-info/requires.txt
+writing top-level names to ../../build/autonomous/autonomous.egg-info/top_level.txt
+reading manifest file '../../build/autonomous/autonomous.egg-info/SOURCES.txt'
+writing manifest file '../../build/autonomous/autonomous.egg-info/SOURCES.txt'
+running build
+running build_py
+copying autonomous/Plane.py -> /home/garrett/Desktop/UAV/build/autonomous/build/lib/autonomous
+running install
+running install_lib
+copying /home/garrett/Desktop/UAV/build/autonomous/build/lib/autonomous/Plane.py -> /home/garrett/Desktop/UAV/install/autonomous/lib/python3.6/site-packages/autonomous
+byte-compiling /home/garrett/Desktop/UAV/install/autonomous/lib/python3.6/site-packages/autonomous/Plane.py to Plane.cpython-36.pyc
+running install_data
+running install_egg_info
+removing '/home/garrett/Desktop/UAV/install/autonomous/lib/python3.6/site-packages/autonomous-0.0.0-py3.6.egg-info' (and everything under it)
+Copying ../../build/autonomous/autonomous.egg-info to /home/garrett/Desktop/UAV/install/autonomous/lib/python3.6/site-packages/autonomous-0.0.0-py3.6.egg-info
+running install_scripts
+Installing talker script to /home/garrett/Desktop/UAV/install/autonomous/lib/autonomous
+writing list of installed files to '/home/garrett/Desktop/UAV/build/autonomous/install.log'
