+++ log/build_2023-05-10_22-42-45/autonomous/streams.log
@@ -1,0 +1,24 @@
+[1.117s] Invoking command in '/home/garrett/Desktop/UAV/src/autonomous': PYTHONPATH=/home/garrett/Desktop/UAV/build/autonomous/prefix_override:/home/garrett/Desktop/UAV/install/autonomous/lib/python3.6/site-packages:${PYTHONPATH} /usr/bin/python3 setup.py egg_info --egg-base ../../build/autonomous build --build-base /home/garrett/Desktop/UAV/build/autonomous/build install --record /home/garrett/Desktop/UAV/build/autonomous/install.log --install-scripts $base/lib/autonomous --single-version-externally-managed
+[1.487s] running egg_info
+[1.489s] writing ../../build/autonomous/autonomous.egg-info/PKG-INFO
+[1.491s] writing dependency_links to ../../build/autonomous/autonomous.egg-info/dependency_links.txt
+[1.492s] writing entry points to ../../build/autonomous/autonomous.egg-info/entry_points.txt
+[1.493s] writing requirements to ../../build/autonomous/autonomous.egg-info/requires.txt
+[1.494s] writing top-level names to ../../build/autonomous/autonomous.egg-info/top_level.txt
+[1.560s] reading manifest file '../../build/autonomous/autonomous.egg-info/SOURCES.txt'
+[1.563s] writing manifest file '../../build/autonomous/autonomous.egg-info/SOURCES.txt'
+[1.564s] running build
+[1.565s] running build_py
+[1.567s] copying autonomous/Plane.py -> /home/garrett/Desktop/UAV/build/autonomous/build/lib/autonomous
+[1.568s] running install
+[1.570s] running install_lib
+[1.572s] copying /home/garrett/Desktop/UAV/build/autonomous/build/lib/autonomous/Plane.py -> /home/garrett/Desktop/UAV/install/autonomous/lib/python3.6/site-packages/autonomous
+[1.573s] byte-compiling /home/garrett/Desktop/UAV/install/autonomous/lib/python3.6/site-packages/autonomous/Plane.py to Plane.cpython-36.pyc
+[1.577s] running install_data
+[1.579s] running install_egg_info
+[1.583s] removing '/home/garrett/Desktop/UAV/install/autonomous/lib/python3.6/site-packages/autonomous-0.0.0-py3.6.egg-info' (and everything under it)
+[1.584s] Copying ../../build/autonomous/autonomous.egg-info to /home/garrett/Desktop/UAV/install/autonomous/lib/python3.6/site-packages/autonomous-0.0.0-py3.6.egg-info
+[1.586s] running install_scripts
+[1.615s] Installing talker script to /home/garrett/Desktop/UAV/install/autonomous/lib/autonomous
+[1.617s] writing list of installed files to '/home/garrett/Desktop/UAV/build/autonomous/install.log'
+[1.650s] Invoked command in '/home/garrett/Desktop/UAV/src/autonomous' returned '0': PYTHONPATH=/home/garrett/Desktop/UAV/build/autonomous/prefix_override:/home/garrett/Desktop/UAV/install/autonomous/lib/python3.6/site-packages:${PYTHONPATH} /usr/bin/python3 setup.py egg_info --egg-base ../../build/autonomous build --build-base /home/garrett/Desktop/UAV/build/autonomous/build install --record /home/garrett/Desktop/UAV/build/autonomous/install.log --install-scripts $base/lib/autonomous --single-version-externally-managed
