+++ log/build_2023-05-10_22-42-45/communications/stdout_stderr.log
@@ -1,0 +1,21 @@
+running egg_info
+writing ../../build/communications/communications.egg-info/PKG-INFO
+writing dependency_links to ../../build/communications/communications.egg-info/dependency_links.txt
+writing entry points to ../../build/communications/communications.egg-info/entry_points.txt
+writing requirements to ../../build/communications/communications.egg-info/requires.txt
+writing top-level names to ../../build/communications/communications.egg-info/top_level.txt
+reading manifest file '../../build/communications/communications.egg-info/SOURCES.txt'
+writing manifest file '../../build/communications/communications.egg-info/SOURCES.txt'
+running build
+running build_py
+copying communications/main.py -> /home/garrett/Desktop/UAV/build/communications/build/lib/communications
+running install
+running install_lib
+copying /home/garrett/Desktop/UAV/build/communications/build/lib/communications/main.py -> /home/garrett/Desktop/UAV/install/communications/lib/python3.6/site-packages/communications
+byte-compiling /home/garrett/Desktop/UAV/install/communications/lib/python3.6/site-packages/communications/main.py to main.cpython-36.pyc
+running install_data
+running install_egg_info
+removing '/home/garrett/Desktop/UAV/install/communications/lib/python3.6/site-packages/communications-0.0.0-py3.6.egg-info' (and everything under it)
+Copying ../../build/communications/communications.egg-info to /home/garrett/Desktop/UAV/install/communications/lib/python3.6/site-packages/communications-0.0.0-py3.6.egg-info
+running install_scripts
+writing list of installed files to '/home/garrett/Desktop/UAV/build/communications/install.log'
