+++ log/build_2023-05-10_22-42-45/communications/streams.log
@@ -1,0 +1,23 @@
+[1.040s] Invoking command in '/home/garrett/Desktop/UAV/src/communications': PYTHONPATH=/home/garrett/Desktop/UAV/build/communications/prefix_override:/home/garrett/Desktop/UAV/install/communications/lib/python3.6/site-packages:${PYTHONPATH} /usr/bin/python3 setup.py egg_info --egg-base ../../build/communications build --build-base /home/garrett/Desktop/UAV/build/communications/build install --record /home/garrett/Desktop/UAV/build/communications/install.log --install-scripts $base/lib/communications --single-version-externally-managed
+[1.404s] running egg_info
+[1.406s] writing ../../build/communications/communications.egg-info/PKG-INFO
+[1.408s] writing dependency_links to ../../build/communications/communications.egg-info/dependency_links.txt
+[1.409s] writing entry points to ../../build/communications/communications.egg-info/entry_points.txt
+[1.410s] writing requirements to ../../build/communications/communications.egg-info/requires.txt
+[1.412s] writing top-level names to ../../build/communications/communications.egg-info/top_level.txt
+[1.480s] reading manifest file '../../build/communications/communications.egg-info/SOURCES.txt'
+[1.483s] writing manifest file '../../build/communications/communications.egg-info/SOURCES.txt'
+[1.485s] running build
+[1.486s] running build_py
+[1.487s] copying communications/main.py -> /home/garrett/Desktop/UAV/build/communications/build/lib/communications
+[1.489s] running install
+[1.490s] running install_lib
+[1.492s] copying /home/garrett/Desktop/UAV/build/communications/build/lib/communications/main.py -> /home/garrett/Desktop/UAV/install/communications/lib/python3.6/site-packages/communications
+[1.494s] byte-compiling /home/garrett/Desktop/UAV/install/communications/lib/python3.6/site-packages/communications/main.py to main.cpython-36.pyc
+[1.496s] running install_data
+[1.497s] running install_egg_info
+[1.502s] removing '/home/garrett/Desktop/UAV/install/communications/lib/python3.6/site-packages/communications-0.0.0-py3.6.egg-info' (and everything under it)
+[1.508s] Copying ../../build/communications/communications.egg-info to /home/garrett/Desktop/UAV/install/communications/lib/python3.6/site-packages/communications-0.0.0-py3.6.egg-info
+[1.511s] running install_scripts
+[1.541s] writing list of installed files to '/home/garrett/Desktop/UAV/build/communications/install.log'
+[1.577s] Invoked command in '/home/garrett/Desktop/UAV/src/communications' returned '0': PYTHONPATH=/home/garrett/Desktop/UAV/build/communications/prefix_override:/home/garrett/Desktop/UAV/install/communications/lib/python3.6/site-packages:${PYTHONPATH} /usr/bin/python3 setup.py egg_info --egg-base ../../build/communications build --build-base /home/garrett/Desktop/UAV/build/communications/build install --record /home/garrett/Desktop/UAV/build/communications/install.log --install-scripts $base/lib/communications --single-version-externally-managed
